SPBLibrary/requirements.txt
asgiref==3.7.2
Django==3.2
djangorestframework==3.12.0
djangorestframework-simplejwt==5.2.2
Pillow==9.5.0
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.2
